Documentation and test code movement around topology
As a part of #8512 I need to have a much more clear understanding of the
topology construct of vector than I presently do. I have an eye toward
accomplishing some of the aims of the Architecture Revisit RFC but would like to
get there in small steps.

This commit does not functionally change vector. It moves some test code around
and adjusts the documentation to be more, I hope, clear about what's what.
