+++ src/main.rs
@@ -12,10 +12,10 @@ fn main() {
 
 #[cfg(windows)]
 pub fn main() {
-    // We need to be able to run vector in User Interactive mode. We first try to run vector
-    // as a service. If we fail, we consider that we are in interactive mode and then fallback
-    // to console mode.
-    // See https://docs.microsoft.com/en-us/dotnet/api/system.environment.userinteractive?redirectedfrom=MSDN&view=netcore-3.1#System_Environment_UserInteractive
+    // We need to be able to run vector in User Interactive mode. We first try
+    // to run vector as a service. If we fail, we consider that we are in
+    // interactive mode and then fallback to console mode.  See
+    // https://docs.microsoft.com/en-us/dotnet/api/system.environment.userinteractive?redirectedfrom=MSDN&view=netcore-3.1#System_Environment_UserInteractive
     vector::vector_windows::run().unwrap_or_else(|_| {
         let app = Application::prepare().unwrap_or_else(|code| {
             std::process::exit(code);
